remove unused calculations service and re-use functions
