+++ src/app/account/account-analysis/account-analysis-dashboard/account-analysis-dashboard.component.ts
@@ -5,9 +5,9 @@ import { ToastNotificationsService } from 'src/app/core-components/toast-notific
 import { AccountAnalysisDbService } from 'src/app/indexedDB/account-analysis-db.service';
 import { AccountdbService } from 'src/app/indexedDB/account-db.service';
 import { DbChangesService } from 'src/app/indexedDB/db-changes.service';
+import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
 import { IdbAccount, IdbAccountAnalysisItem } from 'src/app/models/idb';
 import { SharedDataService } from 'src/app/shared/helper-services/shared-data.service';
-import { AnalysisCalculationsHelperService } from 'src/app/shared/shared-analysis/calculations/analysis-calculations-helper.service';
 
 @Component({
   selector: 'app-account-analysis-dashboard',
@@ -30,7 +30,7 @@ export class AccountAnalysisDashboardComponent implements OnInit {
   yearOptions: Array<number>;
   selectedAccount: IdbAccount;
   constructor(private router: Router, private accountAnalysisDbService: AccountAnalysisDbService, private toastNotificationService: ToastNotificationsService,
-    private accountDbService: AccountdbService, private analysisCalculationsHelperService: AnalysisCalculationsHelperService,
+    private accountDbService: AccountdbService, private utilityMeterDataDbService: UtilityMeterDatadbService,
     private dbChangesService: DbChangesService, private sharedDataService: SharedDataService) { }
 
   ngOnInit(): void {
@@ -39,7 +39,7 @@ export class AccountAnalysisDashboardComponent implements OnInit {
       this.accountAnalysisItems = items;
     });
 
-    this.yearOptions = this.analysisCalculationsHelperService.getYearOptions(true);
+    this.yearOptions = this.utilityMeterDataDbService.getYearOptions(true);
     if (this.yearOptions) {
       this.baselineYearError = this.yearOptions[0] > this.selectedAccount.sustainabilityQuestions.energyReductionBaselineYear
     }
