--- src/app/shared/shared-analysis/calculations/analysis-calculations-helper.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AnalysisCalculationsHelperService } from './analysis-calculations-helper.service';
-
-describe('AnalysisCalculationsHelperService', () => {
-  let service: AnalysisCalculationsHelperService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AnalysisCalculationsHelperService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
